feat(Cars, RacingCarGame, RacingCarGameMain) : RacingCarGame을 자동차들을 관리하는 Cars, 경주를 담당하는 RacingCarGame, 게임 실행 엔진인 RacingCarGameMain으로 분리
